Refactor HID
